Update and Improve use table cost_center instead fee_center
